2009-10-01 17:58 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * INSTALL   * utils/hbmk2/hbmk2.prg   * config/os2/gcc.mk   + config/os2/gccomf.mk     + Added OS/2 GCC OMF support.       This tries to fix resource support in hbmk2 compared to David's       solution. hbmk2 support is completely newly done, .mk support was imported       as is, with little change in actication logic.     ! Fixed to use $(ECHOQUOTE) in some $(ECHO) command parameters.     ; NOTE: Until we don't add autodetection for OMF supporting GCC versions,             it will have to be selected manually using HB_COMPILER=gccomf.             Notice that for GCC 4.x this is obligatory since COFF support is             broken in these versions.             Additional EMX* envvar options are left as an excercise for the             users, I intentionally didn't document these in INSTALL, as INSTALL             servers as Harbour documentation, not 3rd party tools like GCC.
  * config/global.mk
    ! Minor in comment.

  * contrib/hbqt/Makefile
  * contrib/hbqt/hbqts/Makefile
  * contrib/gtqtc/gtqtcs/Makefile
  * contrib/gtqtc/Makefile
    ! Disabled QT dependent parts for msvcia64 targets.
    ; Fails with error:
      ---
      z:\devl\qt\4.5.2\include\qtcore\../../src/corelib/tools/qvector.h(541) : warning C4346: 'QVector<T>::iterator' : dependent name is not a type
              prefix with 'typename' to indicate a type
      z:\devl\qt\4.5.2\include\qtcore\../../src/corelib/tools/qvector.h(541) : error C2143: syntax error : missing ';' before 'QVector<T>::insert'
      z:\devl\qt\4.5.2\include\qtcore\../../src/corelib/tools/qvector.h(541) : error C4430: missing type specifier - int assumed. Note: C++ does not support default-int
      z:\devl\qt\4.5.2\include\qtcore\../../src/corelib/tools/qvector.h(541) : fatal error C1903: unable to recover from previous error(s); stopping compilation
      ---

  * package/winuni/HARBOUR_README_MINGWARM
    * Minor raw update to document newly found --force-local tar switch.

+++ harbour/config/os2/gcc.mk
@@ -8,9 +8,15 @@ else
    HB_CMP := gcc
 endif
 
-OBJ_EXT := .o
-LIB_PREF :=
-LIB_EXT := .a
+ifeq ($(HB_COMPILER),gccomf)
+   OBJ_EXT := .obj
+   LIB_PREF :=
+   LIB_EXT := .lib
+else
+   OBJ_EXT := .o
+   LIB_PREF :=
+   LIB_EXT := .a
+endif
 
 CC := $(HB_CCPATH)$(HB_CCPREFIX)$(HB_CMP)$(HB_CCPOSTFIX)
 CC_IN := -c
@@ -19,6 +25,10 @@ CC_OUT := -o
 CPPFLAGS := -I. -I$(HB_INC_COMPILE)
 CFLAGS :=
 LDFLAGS :=
+
+ifeq ($(HB_COMPILER),gccomf)
+   CFLAGS += -Zomf
+endif
 
 ifneq ($(HB_BUILD_WARN),no)
    CFLAGS += -Wall -W
@@ -62,41 +72,59 @@ LDFLAGS += $(LIBPATHS)
 #
 #endef
 
-# We have to use a script to overcome the AR limit of max 850 characters
-# in commmand line
-define create_library
-   $(if $(wildcard $(subst /,$(DIRSEP),$(LIB_FILE))),@$(RM) $(subst /,$(DIRSEP),$(LIB_FILE)),)
-   @$(ECHO) $(ECHOQUOTE)CREATE $(LIB_DIR)/$@$(ECHOQUOTE) > __lib__.tmp
-   for %i in ( *$(OBJ_EXT) ) do @$(ECHO) $(ECHOQUOTE)ADDMOD %i$(ECHOQUOTE) >> __lib__.tmp
-   @$(ECHO) $(ECHOQUOTE)SAVE$(ECHOQUOTE) >> __lib__.tmp
-   @$(ECHO) $(ECHOQUOTE)END$(ECHOQUOTE) >> __lib__.tmp
-   $(AR) $(ARFLAGS) $(HB_USER_AFLAGS) -M < __lib__.tmp
-endef
+ifeq ($(HB_COMPILER),gccomf)
+   define create_library
+      $(if $(wildcard $(subst /,$(DIRSEP),$(LIB_FILE))),@$(RM) $(subst /,$(DIRSEP),$(LIB_FILE)),)
+      for %i in ( *$(OBJ_EXT) ) do $(AR) $(ARFLAGS) $(HB_USER_AFLAGS) -p128 r $(LIB_DIR)/$@ %i$(ECHOQUOTE)
+   endef
+else
+   # We have to use a script to overcome the AR limit of max 850 characters
+   # in commmand line
+   define create_library
+      $(if $(wildcard $(subst /,$(DIRSEP),$(LIB_FILE))),@$(RM) $(subst /,$(DIRSEP),$(LIB_FILE)),)
+      @$(ECHO) $(ECHOQUOTE)CREATE $(LIB_DIR)/$@$(ECHOQUOTE) > __lib__.tmp
+      for %i in ( *$(OBJ_EXT) ) do @$(ECHO) $(ECHOQUOTE)ADDMOD %i$(ECHOQUOTE) >> __lib__.tmp
+      @$(ECHO) $(ECHOQUOTE)SAVE$(ECHOQUOTE) >> __lib__.tmp
+      @$(ECHO) $(ECHOQUOTE)END$(ECHOQUOTE) >> __lib__.tmp
+      $(AR) $(ARFLAGS) $(HB_USER_AFLAGS) -M < __lib__.tmp
+   endef
+endif
 
 # Under OS/2 || isn't a command separator (inside a shell, that is); correct separator is &
-AR := $(HB_CCPATH)$(HB_CCPREFIX)ar
-ARFLAGS :=
-AR_RULE = $(create_library) $(ARSTRIP) & $(RM) __lib__.tmp
+
+ifeq ($(HB_COMPILER),gccomf)
+   AR := $(HB_CCPATH)$(HB_CCPREFIX)emxomfar
+   ARFLAGS :=
+   AR_RULE = $(create_library) $(ARSTRIP)
+else
+   AR := $(HB_CCPATH)$(HB_CCPREFIX)ar
+   ARFLAGS :=
+   AR_RULE = $(create_library) $(ARSTRIP) & $(RM) __lib__.tmp
+endif
 
 DY := $(CC)
 DFLAGS := -shared $(LIBPATHS)
+ifeq ($(HB_COMPILER),gccomf)
+   DFLAGS += -Zomf
+endif
 DY_OUT := $(LD_OUT)
 DLIBS := $(foreach lib,$(LIBS) $(SYSLIBS),-l$(lib))
 
 # NOTE: The empty line directly before 'endef' HAS TO exist!
 define dyn_object
-   @$(ECHO) $(file)>> __dyn__.tmp
+   @$(ECHO) $(subst $(DIRSEP),/,$(file))>> __dyn__.tmp
    @emxexp $(file)>> __dyn__.def
 
 endef
+
 define create_dynlib
    $(if $(wildcard __dyn__.tmp),@$(RM) __dyn__.tmp,)
    $(if $(wildcard __dyn__.def),@$(RM) __dyn__.def,)
-   @$(ECHO) LIBRARY $(DYNNAME) INITINSTANCE TERMINSTANCE >> __dyn__.def
-   @$(ECHO) PROTMODE >> __dyn__.def
-   @$(ECHO) CODE PRELOAD MOVEABLE DISCARDABLE >> __dyn__.def
-   @$(ECHO) DATA PRELOAD MOVEABLE MULTIPLE NONSHARED >> __dyn__.def
-   @$(ECHO) EXPORTS >> __dyn__.def
+   @$(ECHO) $(ECHOQUOTE)LIBRARY $(DYNNAME) INITINSTANCE TERMINSTANCE$(ECHOQUOTE) >> __dyn__.def
+   @$(ECHO) $(ECHOQUOTE)PROTMODE$(ECHOQUOTE) >> __dyn__.def
+   @$(ECHO) $(ECHOQUOTE)CODE PRELOAD MOVEABLE DISCARDABLE$(ECHOQUOTE) >> __dyn__.def
+   @$(ECHO) $(ECHOQUOTE)DATA PRELOAD MOVEABLE MULTIPLE NONSHARED$(ECHOQUOTE) >> __dyn__.def
+   @$(ECHO) $(ECHOQUOTE)EXPORTS$(ECHOQUOTE) >> __dyn__.def
    $(foreach file,$^,$(dyn_object))
    $(DY) $(DFLAGS) $(HB_USER_DFLAGS) $(DY_OUT)$(DYN_DIR)/$@ @__dyn__.tmp __dyn__.def $(DLIBS) $(DYSTRIP)
    emximp -o $(IMP_FILE) $(DYN_DIR)/$@
